2020-03-08 Roman (stable) - !!!!:DB.LibsDB.{vcaElectroEls,OscadaLibs}: The document of "the Electrical elements library of the user interface" is completed,     all items are revised and the four new items are appended:         "Key, unified (keyUnif)", "Switch, unified (swUnif)", "General state, unified (stGenUnif)", "Area state, unified (stAreaUnif)". - !!!!:SYS,Archive.FSArch,DB.{MySQL,PostgreSQL},DAQ.{JavaLikeCalc,LogicLev,ModBus,Siemens,SoundCard},Protocol.ModBus,         Transport.{SSL,Serial,Sockets},Special.FLibSYS,UI.{QTCfg,VCAEngine,Vision,WebVision}:     !!!!: The source codes significantly cleaned at the building warnings of GCC 9 (CentOs 8) and     !!: The CentOs 8 chroot appended for build OpenSCADA. - DB: The label "Tables lifetime" renamed to "Opened tables lifetime". - !!!!:UI.{Vision,WebVision}: The mirroring support of the primitive FormEl significantly advanced for the shape Arc draw and it generic fills. - !!:UI.Vision: Edition of the primitive FormEl switched to the normal mode, without any transformation (rotation and mirroring).     !: The Arc shapes of the primitive FormEl allowed in connection each other in the fillings buld, like to the sector diagram. - !!: UI.WebVision: The primitive ElFigure source code is significantly cleaned.     Scroll areas of the popupmenu select are set to hidden.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2667 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/ui/VCAEngine/project.cpp
@@ -509,7 +509,7 @@ void Project::pageEnable( const string &pg, bool vl )
     //Process the path
     int pL = 0;
     string pPg, pPath;
-    for(int off = 0; (pPg=TSYS::pathLev(pg,0,false,&off)).size() && off < pg.size(); pL++)
+    for(int off = 0; (pPg=TSYS::pathLev(pg,0,false,&off)).size() && off < (int)pg.size(); pL++)
 	if(pL) pPath += "/"+pPg;
     if(pPg.compare(0,3,"pg_") == 0)	pPg = pPg.substr(3);
 
@@ -764,7 +764,7 @@ void Project::cntrCmdProc( XMLNode *opt )
 
 		// Appending for the properties
 		for(iStPrp = mStProp.begin(); iStPrp != mStProp.end(); iStPrp++)
-		    if(iStPrp->first != "<Styles>" && stlSize() > iStPrp->second.size())
+		    if(iStPrp->first != "<Styles>" && stlSize() > (int)iStPrp->second.size())
 			iStPrp->second.push_back(iStPrp->second[(mStyleIdW>=0)?mStyleIdW:iStPrp->second.size()-1]);
 
 		mStyleIdW = mStProp["<Styles>"].size()-1;
